+++ connector_builder_agents/src/evals/evaluators.py
@@ -11,19 +11,8 @@ from phoenix.evals import OpenAIModel, llm_classify
 
 load_dotenv()
 
-EVAL_MODEL = OpenAIModel(model="gpt-4o")
-
-
-def readiness_eval(input: dict, output: dict) -> int:
-    """Create Phoenix LLM classifier for readiness evaluation."""
-    # Handle case where output is None or missing artifacts
-
-    readiness_report = output.get("artifacts", {}).get("readiness_report", None)
-    if readiness_report is None:
-        print("No readiness report found")
-        return 0
-
-    EVAL_TEMPLATE = """You are evaluating whether a connector readiness test passed or failed.
+READINESS_EVAL_MODEL = OpenAIModel(model="gpt-4o")
+READINESS_EVAL_TEMPLATE = """You are evaluating whether a connector readiness test passed or failed.
 
 A passing report should have:
 - All streams tested successfully (marked with ✅)
@@ -42,12 +31,21 @@ Based on the connector readiness report below, classify whether the test PASSED 
 {readiness_report}
 """
 
+
+def readiness_eval(output: dict) -> int:
+    """Create Phoenix LLM classifier for readiness evaluation. Return 1 if PASSED, 0 if FAILED."""
+
+    readiness_report = output.get("artifacts", {}).get("readiness_report", None)
+    if readiness_report is None:
+        print("No readiness report found")
+        return 0
+
     rails = ["PASSED", "FAILED"]
 
     eval_df = llm_classify(
-        model=EVAL_MODEL,
+        model=READINESS_EVAL_MODEL,
         data=pd.DataFrame([{"readiness_report": readiness_report}]),
-        template=EVAL_TEMPLATE,
+        template=READINESS_EVAL_TEMPLATE,
         rails=rails,
         provide_explanation=True,
     )
@@ -60,8 +58,9 @@ Based on the connector readiness report below, classify whether the test PASSED 
     return score
 
 
-def expected_streams_eval(input: dict, output: dict) -> float:
-    """Evaluate if all expected streams were built."""
+def streams_eval(input: dict, output: dict) -> float:
+    """Evaluate if all expected streams were built. Return the percentage of expected streams that are present in available streams."""
+
     manifest_str = output.get("artifacts", {}).get("manifest", None)
     if manifest_str is None:
         print("No manifest found")
@@ -76,11 +75,10 @@ def expected_streams_eval(input: dict, output: dict) -> float:
     # Get expected streams from the input (dataset row)
     expected_stream_names = ast.literal_eval(input.get("expected_streams", []))
     print(f"Expected stream names: {expected_stream_names}")
-    print(f"Expected stream names type: {type(expected_stream_names)}")
 
     if not expected_stream_names:
         print("No expected streams found")
-        return 0.0  # Avoid division by zero
+        return 0.0
 
     # Calculate the percentage of expected streams that are present in available streams
     matched_streams = set(available_stream_names) & set(expected_stream_names)
