+++ poe_tasks.toml
@@ -45,8 +45,7 @@ docs-generate.help = "Generate API documentation"
 # Agentic Connector Build Testing
 build-connector = { cmd = "uv run --project=connector_builder_agents python -m connector_builder_agents.src.cli", help = "Run the MCP agent" }
 build-connector-interactive = { cmd = "uv run --project=connector_builder_agents python -m connector_builder_agents.src.cli --interactive", help = "Run the MCP agent interactively (user is manager)" }
-run-evals = { cmd = "uv run --project=connector_builder_agents python -m connector_builder_agents.src.eval_cli", help = "Run connector builder evaluations from YAML config" }
-run-evals-phoenix = { cmd = "uv run --project=connector_builder_agents python -m connector_builder_agents.src.evals_phoenix", help = "Run connector builder evaluations from YAML config" }
+run-evals = { cmd = "uv run --project=connector_builder_agents python -m connector_builder_agents.src.evals.phoenix_run", help = "Run connector builder evaluations from YAML config" }
 
 
 # Codex Tests
